requirements.txt
asgiref==3.4.1
da-vinci==0.3.0
Django==3.2.5
django-jalali==4.3.0
django-mathfilters==1.0.0
django-render-partial==0.4
django-thumbnails==0.6.0
djangorestframework==3.12.4
jdatetime==3.6.2
Pillow==8.3.1
pytz==2021.1
shortuuid==1.0.1
sqlparse==0.4.1
